+++ src/components/UserCards/ConnectCard.tsx
@@ -1,0 +1,8 @@
+import { PublicUser } from "../../utils/types";
+
+interface ConnectCardProps {
+  user: PublicUser;
+}
+export const ConnectCard = (props: ConnectCardProps): JSX.Element => {
+  return <div>Whattttt</div>;
+};
